+++ hera_librarian/base_store.py
@@ -18,13 +18,15 @@ BaseStore
 
 import subprocess
 import os.path
+import time
+import warnings
 
 from . import RPCError
 
 NUM_RSYNC_TRIES = 6
 
 
-class BaseStore (object):
+class BaseStore(object):
     """Note that the Librarian server code subclasses this class, so do not change
     its structure without making sure that you're not breaking it.
 
@@ -101,19 +103,24 @@ class BaseStore (object):
         stdin.close()
         return proc
 
-    # Modifications of the store host. These should always be paired with
-    # appropriate modifications of the Librarian server database, either
-    # through an RPC call (if you're a client) or a direct change (if you're
-    # the server).
+    def _rsync_transfer(self, local_path, store_path):
+        """Copy a file to a particular path using rsync.
 
-    def copy_to_store(self, local_path, store_path):
-        """Rsync a file to a particular path in the store.
+        Parameters
+        ----------
+        local_path : str
+            Path to local file to be copied.
+        store_path : str
+            Path to store file at on destination host.
 
-        You should not copy files directly to their intended destinations. You
-        should use the Librarian `prepare_upload` RPC call to get a staging
-        directory and copy your files there; then use the `complete_upload`
-        RPC call to tell the Librarian that you're done.
+        Returns
+        -------
+        None
 
+        Raises
+        ------
+        RPCError
+            Raised if rsync transfer does not complete successfully.
         """
         # Rsync will nest directories in a way that we don't want if we don't
         # end their names with "/", but it will error if we end a file name
@@ -134,8 +141,10 @@ class BaseStore (object):
         argv = [
             'rsync',
             '-aP',
-            '-e', 'ssh -c aes128-ctr -o BatchMode=yes -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no',
-            local_path + local_suffix, '%s:%s' % (self.ssh_host, self._path(store_path))
+            '-e',
+            'ssh -c aes128-ctr -o BatchMode=yes -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no',
+            local_path + local_suffix,
+            '%s:%s' % (self.ssh_host, self._path(store_path))
         ]
         success = False
 
@@ -150,6 +159,193 @@ class BaseStore (object):
 
         if not success:
             raise RPCError(argv, 'exit code %d; output:\n\n%r' % (proc.returncode, output))
+
+    def _globus_transfer(
+        self,
+        local_path,
+        store_path,
+        client_id,
+        transfer_token,
+        source_endpoint_id,
+        destination_endpoint_id,
+        host_path,
+    ):
+        """Copy a file to a particular path using globus.
+
+        For further information on using the Python API for Globus, refer to the
+        Globus SDK docs: https://globus-sdk-python.readthedocs.io/en/stable/
+
+        Parameters
+        ----------
+        local_path : str
+            Path to local file to be copied.
+        store_path : str
+            Path to store file at on destination host. Note that if the target
+            endpoint is a "shared endpoint", this may be relative to the globus
+            root directory rather than the filesystem root.
+        client_id : str
+            The globus client ID to use for the transfer.
+        transfer_token : str
+            The globus transfer token to use for the transfer.
+        source_endpoint_id : str
+            The globus endpoint ID of the source store.
+        destination_endpoint_id : str
+            The globus endpoint ID of the destination store.
+        host_path : str, optional
+            The `host_path` of the globus store. When using shared endpoints,
+            this is the root directory presented to the client.
+
+        Returns
+        -------
+        None
+
+        Raises
+        ------
+        RPCError
+            Raised if submission of transfer to globus does not complete
+            successfully or if globus_sdk package is not installed.
+        """
+        try:
+            import globus_sdk
+            from globus_sdk.exc import AuthAPIError
+        except ModuleNotFoundError:
+            raise RPCError(
+                "globus_sdk import",
+                "The globus_sdk package must be installed for globus "
+                "functionality. Please `pip install globus_sdk` and try again."
+            )
+
+        # check that both endpoint IDs have been specified
+        if source_endpoint_id is None or destination_endpoint_id is None:
+            raise RPCError(
+                "globus endpoint check",
+                "Both source_endpoint_id and destination_endpoint_id must be "
+                "specified to initiate globus transfer."
+            )
+
+        # make globus transfer client
+        client = globus_sdk.NativeAppAuthClient(client_id)
+        try:
+            authorizer = globus_sdk.RefreshTokenAuthorizer(transfer_token, client)
+        except AuthAPIError:
+            raise RPCError(
+                "globus authorization",
+                "globus authentication failed. Please check client_id and "
+                "authentication token and try again."
+            )
+        tc = globus_sdk.TransferClient(authorizer=authorizer)
+
+        # make a new data transfer object
+        basename = os.path.basename(os.path.normpath(local_path))
+        tdata = globus_sdk.TransferData(
+            tc,
+            source_endpoint_id,
+            destination_endpoint_id,
+            sync_level="checksum",
+            label=basename,
+            notify_on_succeeded=False,
+            notify_on_failed=True,
+            notify_on_inactive=True,
+        )
+
+        # format the path correctly
+        store_path = self._path(store_path)
+        if host_path is not None:
+            if store_path.startswith(host_path):
+                store_path = store_path[len(host_path):]
+                # trim off leading "/" if present
+                if store_path.startswith("/"):
+                    store_path = store_path.lstrip("/")
+
+        # add data to be transferred
+        tdata.add_item(local_path, store_path)
+
+        # initiate transfer
+        transfer_result = tc.submit_transfer(tdata)
+
+        # get task_id and query status until it finishes
+        task_id = transfer_result["task_id"]
+        while True:
+            task = tc.get_task(task_id)
+            if task["status"] == "SUCCEEDED":
+                return
+            elif task["status"] == "FAILED":
+                # get the events associated with this transfer to help with debugging
+                events = tc.task_event_list(task_id)
+                error_string = ""
+                for event in events:
+                    error_string.append(event["time"] + ": " + event["description"] + "\n")
+                raise RPCError("globus transfer", 'events:\n\n%r' % (error_string))
+            else:  # task is "ACTIVE"
+                time.sleep(5)
+
+    # Modifications of the store host. These should always be paired with
+    # appropriate modifications of the Librarian server database, either
+    # through an RPC call (if you're a client) or a direct change (if you're
+    # the server).
+
+    def copy_to_store(
+        self,
+        local_path,
+        store_path,
+        try_globus=False,
+        client_id=None,
+        transfer_token=None,
+        source_endpoint_id=None,
+        destination_endpoint_id=None,
+        host_path=None,
+    ):
+        """Transfer a file to a particular path in the store.
+
+        You should not copy files directly to their intended destinations. You
+        should use the Librarian `prepare_upload` RPC call to get a staging
+        directory and copy your files there; then use the `complete_upload`
+        RPC call to tell the Librarian that you're done.
+
+        Parameters
+        ----------
+        local_path : str
+            Path to the local file to upload.
+        store_path : str
+            Path to store file at on destination host.
+        try_globus : bool, optional
+            Whether to try to use globus to transfer. If False, or if globus
+            fails, automatically fall back on rsync.
+        client_id : str, optional
+            The globus client ID to use for the transfer.
+        transfer_token : str, optional
+            The globus transfer token to use for the transfer.
+        source_endpoint_id : str, optional
+            The globus endpoint ID of the source store.
+        destination_endpoint_id : str, optional
+            The globus endpoint ID of the destination store.
+        host_path : str, optional
+            The `host_path` of the globus store. When using shared endpoints,
+            this is the root directory presented to the client. Note that this
+            may be different from the `path_prefix` for a given store.
+
+        Returns
+        -------
+        None
+        """
+        if try_globus:
+            try:
+                self._globus_transfer(
+                    local_path,
+                    store_path,
+                    client_id,
+                    transfer_token,
+                    source_endpoint_id,
+                    destination_endpoint_id,
+                    host_path,
+                )
+            except RPCError as e:
+                # something went wrong with globus--fall back on rsync
+                print("Globus transfer failed: {}\nFalling back on rsync...".format(e))
+                self._rsync_transfer(local_path, store_path)
+        else:
+            # use rsync from the get-go
+            self._rsync_transfer(local_path, store_path)
 
     def _chmod(self, store_path, modespec):
         """Change Unix permissions on a path in the store.
@@ -208,8 +404,10 @@ class BaseStore (object):
         else:
             piece = ''
 
-        return self._ssh_slurp("mkdir -p '%s' && chmod u+w '%s' && mv -nT '%s' '%s' && test ! -e '%s'%s" %
-                               (self._path(dest_parent), ssp, ssp, dsp, ssp, piece))
+        return self._ssh_slurp(
+            "mkdir -p '%s' && chmod u+w '%s' && mv -nT '%s' '%s' && test ! -e '%s'%s" %
+            (self._path(dest_parent), ssp, ssp, dsp, ssp, piece)
+        )
 
     def _delete(self, store_path, chmod_before=False):
         """Delete a path from the store.
@@ -254,8 +452,10 @@ class BaseStore (object):
 
         """
         import json
-        text = self._ssh_slurp("python -c \'import hera_librarian.utils as u; u.print_info_for_path(\"%s\")\'"
-                               % (self._path(storepath)))
+        text = self._ssh_slurp(
+            "python -c \'import hera_librarian.utils as u; u.print_info_for_path(\"%s\")\'"
+            % (self._path(storepath))
+        )
         return json.loads(text)
 
     _cached_space_info = None
@@ -319,19 +519,65 @@ class BaseStore (object):
         info = self.get_space_info()
         return 100. * info['used'] / (info['total'])
 
-    def upload_file_to_other_librarian(self, conn_name, rec_info, local_store_path,
-                                       remote_store_path=None, known_staging_store=None,
-                                       known_staging_subdir=None):
-        """Fire off an rsync process on the store that will upload a given file to a
-        different Librarian. This function will SSH into the store host, from
-        which it will launch an rsync, and it will not return until everything
-        is done! This means that in the real world it may not return for
-        hours, and it will not infrequently raise an exception.
+    def upload_file_to_other_librarian(
+        self,
+        conn_name,
+        rec_info,
+        local_store_path,
+        remote_store_path=None,
+        known_staging_store=None,
+        known_staging_subdir=None,
+        use_globus=False,
+        client_id=None,
+        transfer_token=None,
+        source_endpoint_id=None,
+    ):
+        """Upload a given file to a different Librarian.
 
-        TODO: there is no progress tracking; we just block and eventually
-        return some textual output from the rsync-within-SSH. This is far from
-        ideal.
+        This function will SSH into the store host, from which it will launch an
+        rsync or globus transfer, and it will not return until everything is
+        done! This means that in the real world it may not return for hours, and
+        it will not infrequently raise an exception.
 
+        TODO: there is no internal progress tracking; we just block and
+        eventually return some textual output from the rsync-within-SSH. This is
+        far from ideal. Globus offers some information about transfer progress,
+        but we do not (yet) have automated ways of retrieving it.
+
+        Parameters
+        ----------
+        conn_name : str
+            The name of the external librarian to upload a file to.
+        rec_info : dict
+            A dictionary containing information about the file records being
+            transferred. These are needed by the receiving librarian.
+        local_store_path : str
+            The full path to the file in the local store.
+        remote_store_path : str, optional
+            The full path to the file destination in the remote store. If not
+            specified, it will default to be the same as local_store_path.
+        known_staging_store : str, optional
+            The store corresponding to the already-uploaded file. Must be specified
+            if `known_staging_subdir` is specified.
+        known_staging_subdir : str, optional
+            The target directory corresponding to the already-uploaded file. Must by
+            specified if `known_staging_store` is specified.
+        use_globus : bool, optional
+            Whether to try to use globus to transfer. If False, or if globus
+            fails, automatically fall back on rsync.
+        client_id : str, optional
+            The globus client ID to use for the transfer.
+        transfer_token : str, optional
+            The globus transfer token to use for the transfer.
+        source_endpoint_id : str, optional
+            The globus endpoint ID of the source store. May be omitted, in which
+            case we assume it is a "personal" (as opposed to public) client.
+
+        Returns
+        -------
+        bytes
+            The output of the `_ssh_slurp` command to log into the store and
+            launch the librarian upload.
         """
         if remote_store_path is None:
             remote_store_path = local_store_path
@@ -349,6 +595,13 @@ class BaseStore (object):
 
         command = 'librarian upload --meta=json-stdin%s %s %s %s' % (
             pre_staged_arg, conn_name, self._path(local_store_path), remote_store_path)
+        # optional globus additions to the command
+        if use_globus:
+            command += f" --use_globus --client_id={client_id} --transfer_token={transfer_token}"
+            if source_endpoint_id is not None:
+                command += f" --source_endpoint_id={source_endpoint_id}"
+
+        # actually run the command
         return self._ssh_slurp(command, input=rec_text.encode("utf-8"))
 
     def upload_file_to_local_store(self, local_store_path, dest_store, dest_rel):