+++ librarian_server/store.py
@@ -36,7 +36,7 @@ from .dbutil import NotNull, SQLAlchemyError
 from .webutil import ServerError, json_api, login_required, optional_arg, required_arg
 
 
-class Store (db.Model, BaseStore):
+class Store(db.Model, BaseStore):
     """A Store is a computer with a disk where we can store data. Several of the
     things we keep track of regarding stores are essentially configuration
     items; but we also keep track of the machine's availability, which is
@@ -71,11 +71,11 @@ class Store (db.Model, BaseStore):
 
     def convert_to_base_object(self):
         """Asynchronous store operations are run on worker threads, which means that
-        they're not allowed to access the database. But we'd like to be able
-        to pass Store references around and reuse the functionality
-        implemented in the `hera_librarian.store.Store` class. So we have this
-        helper function that converts this fancy, database-enabled object into
-        a simpler one that can be passed to other threads and so on.
+        they're not allowed to access the database. But we'd like to be able to
+        pass Store references around and reuse the functionality implemented in
+        the `hera_librarian.base_store.BaseStore` class. So we have this helper
+        function that converts this fancy, database-enabled object into a
+        simpler one that can be passed to other threads and so on.
 
         """
         return BaseStore(self.name, self.path_prefix, self.ssh_host)
@@ -429,19 +429,66 @@ def register_instances(args, sourcename=None):
 from . import bgtasks
 
 
-class UploaderTask (bgtasks.BackgroundTask):
+class UploaderTask(bgtasks.BackgroundTask):
     """Object that manages the task of copying a file to another Librarian.
 
-    `remote_store_path` may be None, in which case we will request the same
-    "store path" as the file was used in this Librarian by whichever
-    FileInstance we happen to have located.
+    If `known_staging_store` and `known_staging_subdir` are not None, the copy
+    will be launched assuming that files have already been staged at a known
+    location at the final destination. This is useful if files have been
+    copied from one Librarian site to another outside of the Librarian
+    framework.
 
+    Parameters
+    ----------
+    store : BaseStore object
+        A BaseStore object corresponding to the originating store.
+    conn_name : str
+        The name of the connection to use, as defined in ~/.hl_client.cfg.
+    rec_info : dict
+        A dictionary containing database information for the file to be
+        transferred.
+    store_path : str
+        The full path to the file in the local store.
+    remote_store_path : str, optional
+        The path to place the file in the destination store. This may be None,
+        in which case we will request the same "store path" as the FileInstance
+        in this Librarian.
+    standing_order_name : str, optional
+        The standing order corresponding to this upload task.
+    known_staging_store : str, optional
+        The store corresponding to the already-uploaded file. Must be specified
+        if `known_staging_subdir` is specified.
+    known_staging_subdir : str, optional
+        The target directory corresponding to the already-uploaded file. Must by
+        specified if `known_staging_store` is specified.
+    use_globus : bool, optional
+        Specify whether to try to use globus to transfer files.
+    client_id : str, optional
+        The globus client ID to use for the transfer.
+    transfer_token : str, optional
+        The globus transfer token to use for the transfer.
+    source_endpoint_id : str, optional
+        The globus endpoint ID of the source store. May be omitted, in which
+        case we assume it is a "personal" (as opposed to public) client.
     """
     t_start = None
     t_finish = None
 
-    def __init__(self, store, conn_name, rec_info, store_path, remote_store_path, standing_order_name=None,
-                 known_staging_store=None, known_staging_subdir=None):
+    def __init__(
+        self,
+        store,
+        conn_name,
+        rec_info,
+        store_path,
+        remote_store_path,
+        standing_order_name=None,
+        known_staging_store=None,
+        known_staging_subdir=None,
+        use_globus=False,
+        client_id=None,
+        transfer_token=None,
+        source_endpoint_id=None,
+    ):
         self.store = store
         self.conn_name = conn_name
         self.rec_info = rec_info
@@ -450,6 +497,10 @@ class UploaderTask (bgtasks.BackgroundTask):
         self.standing_order_name = standing_order_name
         self.known_staging_store = known_staging_store
         self.known_staging_subdir = known_staging_subdir
+        self.use_globus = use_globus
+        self.client_id = client_id
+        self.transfer_token = transfer_token
+        self.source_endpoint_id = source_endpoint_id
 
         self.desc = 'upload %s:%s to %s:%s' % (store.name, store_path,
                                                conn_name, remote_store_path or '<any>')
@@ -461,10 +512,17 @@ class UploaderTask (bgtasks.BackgroundTask):
         import time
         self.t_start = time.time()
         self.store.upload_file_to_other_librarian(
-            self.conn_name, self.rec_info,
-            self.store_path, self.remote_store_path,
+            self.conn_name,
+            self.rec_info,
+            self.store_path,
+            self.remote_store_path,
             known_staging_store=self.known_staging_store,
-            known_staging_subdir=self.known_staging_subdir)
+            known_staging_subdir=self.known_staging_subdir,
+            use_globus=self.use_globus,
+            client_id=self.client_id,
+            transfer_token=self.transfer_token,
+            source_endpoint_id=self.source_endpoint_id,
+        )
         self.t_finish = time.time()
 
     def wrapup_function(self, retval, exc):
@@ -520,9 +578,15 @@ class UploaderTask (bgtasks.BackgroundTask):
             raise ServerError('failed to commit completion events to database')
 
 
-def launch_copy_by_file_name(file_name, connection_name, remote_store_path=None,
-                             standing_order_name=None, no_instance='raise',
-                             known_staging_store=None, known_staging_subdir=None):
+def launch_copy_by_file_name(
+    file_name,
+    connection_name,
+    remote_store_path=None,
+    standing_order_name=None,
+    no_instance='raise',
+    known_staging_store=None,
+    known_staging_subdir=None,
+):
     """Launch a copy of a file to a remote Librarian.
 
     A ServerError will be raised if no instance of the file is available.
@@ -565,18 +629,44 @@ def launch_copy_by_file_name(file_name, connection_name, remote_store_path=None,
     from .misc import gather_records
     rec_info = gather_records(file)
 
+    # Figure out if we should try to use globus or not
+    if app.config["use_globus"]:
+        source_endpoint_id = app.config.get("globus_endpoint_id", None)
+        try:
+            client_id = app.config["globus_client_id"]
+            transfer_token = app.config["globus_transfer_token"]
+            use_globus = True
+        except KeyError:
+            client_id = None
+            transfer_token = None
+            use_globus = False
+    else:
+        use_globus = False
+        client_id = None
+        transfer_token = None
+        source_endpoint_id = None
+
     # Launch the background task. We need to convert the Store to a base object since
     # the background task can't access the database.
-
     basestore = inst.store_object.convert_to_base_object()
-    bgtasks.submit_background_task(UploaderTask(
-        basestore, connection_name, rec_info, inst.store_path,
-        remote_store_path, standing_order_name,
-        known_staging_store=known_staging_store,
-        known_staging_subdir=known_staging_subdir))
+    bgtasks.submit_background_task(
+        UploaderTask(
+            basestore,
+            connection_name,
+            rec_info,
+            inst.store_path,
+            remote_store_path,
+            standing_order_name,
+            known_staging_store=known_staging_store,
+            known_staging_subdir=known_staging_subdir,
+            use_globus=use_globus,
+            client_id=client_id,
+            transfer_token=transfer_token,
+            source_endpoint_id=source_endpoint_id,
+        )
+    )
 
     # Remember that we launched this copy.
-
     db.session.add(file.make_copy_launched_event(connection_name, remote_store_path))
 
     try: